Refactor common codegen code
Extract the main code generator loop, the register management methods, and some debug info code.
